requirements.txt
django
pymysql
requests
pyyaml
django-cors-headers
pytz
djangorestframework
djangorestframework-simplejwt
pillow